core seems fine, except Domains

+++ oscar-cbls/src/main/scala/oscar/cbls/business/routing/invariants/capa/ForwardCumulativeInteger5DimensionOnVehicle.scala
@@ -143,7 +143,7 @@ class ForwardCumulativeInteger5DimensionOnVehicle(routes:ChangingSeqValue,
 
   for(i <- lastPointOfVehicle) i.setDefiningInvariant(this)
 
-  override def notifyIntChanged(v: ChangingIntValue, id: Long, OldVal: Long, NewVal: Long){
+  override def notifyIntChanged(v: ChangingIntValue, id: Int, OldVal: Long, NewVal: Long) {
     toUpdateZonesAndVehicleStartAfter match {
       case None => ;
       case Some((toUpdateZones,vehicleLocation)) =>
