+++ oscar-cbls/src/main/scala/oscar/cbls/lib/invariant/logic/BinPackingLoad.scala
@@ -33,7 +33,7 @@ case class BinPackingLoad(items: Array[IntValue], itemsizes: Array[Long]) extend
   }
 
   @inline
-  override def notifyIntChanged(v: ChangingIntValue, id: Long, OldVal: Long, NewVal: Long): Unit = {
+  override def notifyIntChanged(v: ChangingIntValue, id: Int, OldVal: Long, NewVal: Long): Unit = {
     binLoad(OldVal) :-= itemsizes(id)
     binLoad(NewVal) :+= itemsizes(id)
   }
