+++ oscar-cbls/src/main/scala/oscar/cbls/lib/invariant/numeric/RoundUp.scala
@@ -125,7 +125,7 @@ case class RoundUpCustom(from: IntValue, duration: IntValue, forbiddenZones: Lis
   this := roundup()
 
   @inline
-  override def notifyIntChanged(v: ChangingIntValue, id:Long, OldVal: Long, NewVal: Long) {
+  override def notifyIntChanged(v: ChangingIntValue, id: Int, OldVal: Long, NewVal: Long) {
     scheduleForPropagation()
   }
