+++ oscar-cbls/src/main/scala/oscar/cbls/lib/invariant/seq/OccurrencesOf.scala
@@ -38,7 +38,7 @@ case class OccurrencesOf(v: SeqValue, a:IntValue)
     scheduleForPropagation()
   }
 
-  override def notifyIntChanged(v : ChangingIntValue, id : Long, OldVal : Long, NewVal : Long){
+  override def notifyIntChanged(v: ChangingIntValue, id: Int, OldVal: Long, NewVal: Long) {
     scheduleForPropagation()
   }
